+++ src/utils.h
@@ -32,6 +32,6 @@ void pi_controller(struct_pi_controller_state *pi_controller_state);
 void pi_controller_reset(struct_pi_controller_state *pi_controller);
 void crc16(uint8_t ui8_data, uint16_t* ui16_crc);
 void ui8_limit_max(uint8_t *ui8_p_value, uint8_t ui8_max_value);
-void ui16_limit_max(uint8_t *ui16_p_value, uint8_t ui16_max_value);
+void ui16_limit_max(uint16_t *ui16_p_value, uint16_t ui16_max_value);
 
 #endif /* _UTILS_H */
